Refactor directories; fix setup.py and README
