--- README
@@ -1,21 +1,0 @@
-March 7, 2016
-
-To run:
-% python spectator.py targets
-where targets.list is a file to be read in with
-first column = flag (0,1) if target is to be used,
-second column = target/directory name"  
-
-This will run both scrape_headers (which generates tables of information describing the sample) and quick_look, which generates demographics plots and plots of spectra
-
-Required packages:
- - astropy: http://www.astropy.org/
- - fitsio: https://pypi.python.org/pypi/fitsio/
-   (soon to be optional if you have astropy)
- - matplotlib
- - Assumes Python 2.
-
-
-Known issues:
-
-Currently this only works with data taken with the Cosmic Origins Spectrograph (COS) aboard the Hubble Space Telescope, and many of the peculiarities, both in plotting and scraping the headers for information, are particular to COS. That said, it does work for _all_ COS data, both in the FUV and NUV, publicly available as of February 2016.
